refactor : 나머지 컨트롤러들도 어노테이션 기반으로 변경
