+++ src/main/java/core/mvc/asis/LegacyRequestMapping.java
@@ -15,15 +15,15 @@ public class LegacyRequestMapping implements HandlerMapping {
 
     void initMapping() {
 //        mappings.put("/", new HomeController());
-        mappings.put("/users/form", new ForwardController("/user/form.jsp"));
-        mappings.put("/users/loginForm", new ForwardController("/user/login.jsp"));
-        mappings.put("/users", new ListUserController());
-        mappings.put("/users/login", new LoginController());
-        mappings.put("/users/profile", new ProfileController());
-        mappings.put("/users/logout", new LogoutController());
+//        mappings.put("/users/form", new ForwardController("/user/form.jsp"));
+//        mappings.put("/users/loginForm", new ForwardController("/user/login.jsp"));
+//        mappings.put("/users", new ListUserController());
+//        mappings.put("/users/login", new LoginController());
+//        mappings.put("/users/profile", new ProfileController());
+//        mappings.put("/users/logout", new LogoutController());
 //        mappings.put("/users/create", new CreateUserController());
-        mappings.put("/users/updateForm", new UpdateFormUserController());
-        mappings.put("/users/update", new UpdateUserController());
+//        mappings.put("/users/updateForm", new UpdateFormUserController());
+//        mappings.put("/users/update", new UpdateUserController());
 
         logger.info("Initialized Request Mapping!");
         mappings.keySet().forEach(path -> {
