+++ src/main/java/next/controller/ListUserController.java
@@ -1,13 +1,15 @@
 package next.controller;
 
+import core.annotation.web.RequestMapping;
 import core.db.DataBase;
 import core.mvc.asis.Controller;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-public class ListUserController implements Controller {
-    @Override
+@core.annotation.web.Controller
+public class ListUserController {
+    @RequestMapping(value = "/users")
     public String execute(HttpServletRequest req, HttpServletResponse resp) throws Exception {
         if (!UserSessionUtils.isLogined(req.getSession())) {
             return "redirect:/users/loginForm";
