+++ src/main/java/next/controller/LogoutController.java
@@ -1,13 +1,15 @@
 package next.controller;
 
+import core.annotation.web.RequestMapping;
 import core.mvc.asis.Controller;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
-public class LogoutController implements Controller {
-    @Override
+@core.annotation.web.Controller
+public class LogoutController {
+    @RequestMapping(value = "/users/logout")
     public String execute(HttpServletRequest req, HttpServletResponse resp) throws Exception {
         HttpSession session = req.getSession();
         session.removeAttribute(UserSessionUtils.USER_SESSION_KEY);
